fast preview script
added quick utility for animating in blender.   this will playback (or preview) the current range from the beginning and when pressed again, will return the current frame to the frame before preview.  This is most usefull when refining a pose and wanting to see how it feels in motion.  I recommend binding it to the spacebar for animators (and moving the search to another key, like command+s)